refactor registration
toaster, validation...
