+++ src/pages/About/About.scss
@@ -1,4 +1,4 @@
-@import "src/constants";
+@import "../../constants";
 
 .About {
   display: grid;
