+++ src/pages/Registration/Registration.scss
